# schema/openapi.yaml
openapi: 3.0.3
info:
  title: ivaCAM API
  version: 0.0.1
  description: |
    JSON contract for ivaCAM core operations.
    Implementation-agnostic: served by the Stage-1 Python FastAPI bridge,
    the Rust axum server, Tauri command bindings, and the WASM browser
    bindings — same shape everywhere.
  license:
    name: GPL-3.0-or-later
servers:
- url: http://localhost:8765
  description: local Stage-1 FastAPI bridge
paths:
...
  /import:
    post:
      summary: Parse a vector file into ivaCAM geometry
      operationId: import
      description: |
        Accepts the bytes of a DXF, SVG, HPGL, NGC, STL or other supported
        format and returns the flattened geometry: layers, segments
        (LINE/ARC primitives with bulge for arcs), bounding box, and unit
        scale.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - file
              properties:
                file:
                  type: string
                  format: binary
                format:
                  type: string
                  description: Optional explicit format hint (dxf|svg|hpgl|ngc|stl). Inferred from filename otherwise.
      responses:
        '200':
          description: Parsed geometry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /generate:
    post:
      summary: Generate gcode + 3D toolpath preview from imported geometry
      operationId: generate
      description: |
        Stub for Stage-1; full setup-tree input lands in d0d follow-ups.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
      responses:
        '200':
          description: Generated output
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /text:
    post:
      summary: Render TTF text into geometry segments
      operationId: renderText
      description: |
        Tessellates a TTF/OTF font + a string into the project's standard
        Segment shape. Used by the Add-Text dialog. The response also
        carries a `single_line` flag derived by `is_single_line_font` so
        the dialog can warn when a filled-outline font is paired with the
        Engraving style.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenderTextRequest'
      responses:
        '200':
          description: Rendered text geometry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderTextResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /helix-radius:
    post:
      summary: Preview the auto-fit helix radius for a pocket selection
      operationId: computeHelixRadius
      description: |
        Runs the same inscribed-circle search the gcode generator does at
        run time, so the OpPropertiesPanel can display the detected radius
        (or a short fallback reason) before the user clicks Generate.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HelixRadiusRequest'
      responses:
        '200':
          description: Detected helix radius or a fallback reason
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelixRadiusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
components:
  schemas:
...
    BBox:
      properties:
        max_x:
          format: double
          type: number
        max_y:
          format: double
          type: number
        min_x:
          format: double
          type: number
        min_y:
          format: double
          type: number
      required:
      - max_x
      - max_y
      - min_x
      - min_y
      type: object
...
    Error:
      properties:
        details: true
        error:
          type: string
      required:
      - error
      type: object
...
    GenerateRequest:
      properties:
        post_processor:
          anyOf:
          - $ref: '#/components/schemas/PostProcessorKind'
          - type: 'null'
        project:
          allOf:
          - $ref: '#/components/schemas/Project'
          description: The full project — geometry + machine + tools + operations + tabs.
      required:
      - project
      type: object
    GenerateResponse:
      properties:
        gcode:
          type: string
        gcode_index:
          $ref: '#/components/schemas/GcodeIndex'
        regions:
          description: 'Filled-area preview for Pocket ops: the actual region the cutter will machine, computed via the per-op `SourceCombine` mode (Auto by default — outer + inner = annulus). The frontend paints these as translucent fills so the user sees what they''re cutting before reading the toolpath. Empty for non-Pocket ops.'
          items:
            $ref: '#/components/schemas/RegionPreview'
          type: array
        stats:
          $ref: '#/components/schemas/PipelineStats'
        time_estimate:
          allOf:
          - $ref: '#/components/schemas/TimeEstimate'
          description: Acceleration- and jerk-aware program-time estimate. See [`crate::sim::timing`] for the integrator. The total accounts for motion under the trapezoidal profile, tool-change time (`MachineConfig.toolchange_s` × number of M6s), and per-tool spindle pauses summed across used tools.
        toolpath:
          items:
            $ref: '#/components/schemas/ToolpathSegment'
          type: array
        warnings:
          description: Non-fatal warnings raised during planning — mostly tool-fit problems (cutter doesn't fit the geometry, kind mismatch, …). The frontend surfaces these in the operations list status badge and a sidebar list; the gcode is still emitted.
          items:
            $ref: '#/components/schemas/PipelineWarning'
          type: array
      required:
      - gcode
      - gcode_index
      - stats
      - time_estimate
      - toolpath
      type: object
...
    HelixRadiusRequest:
      description: Cross-transport request for the helix auto-fit preview. The frontend sends imported segments + the operation's selected object ids + the active tool diameter; the backend walks the same chaining/combine pipeline the gcode generator does and returns the largest inscribed circle that fits inside the resulting pocket boundary.
      properties:
        object_ids:
          description: Object ids from the import that should participate in the pocket boundary computation. Same shape as `OpSource::Objects.ids`. Empty = "use all segments as one region".
          items:
            format: uint32
            minimum: 0.0
            type: integer
          type: array
        segments:
          description: Source segments to derive the closed-pocket boundary from.
          items:
            $ref: '#/components/schemas/Segment'
          type: array
        tool_diameter_mm:
          format: double
          type: number
      required:
      - object_ids
      - segments
      - tool_diameter_mm
      type: object
    HelixRadiusResponse:
      properties:
        fallback_reason:
          description: Short reason string when None (so the UI can show "Pocket too small for tool" instead of just an empty Auto state).
          type:
          - string
          - 'null'
        radius_mm:
          description: Some(r) when an inscribed circle fits; None when the auto path would fall back to Ramp/Direct.
          format: double
          type:
          - number
          - 'null'
      type: object
...
    ImportResponse:
      properties:
        bbox:
          $ref: '#/components/schemas/BBox'
        filename:
          type: string
        format:
          type: string
        layers:
          items:
            $ref: '#/components/schemas/Layer'
          type: array
        object_meta:
          default: []
          description: Object metadata (closed flag, layer, color, bbox) keyed by the same 1-based id — `object_meta[id - 1]` is the entry for object `id`.
          items:
            $ref: '#/components/schemas/ImportedObject'
          type: array
        objects:
          default: []
          description: Per-segment object id from the chaining pass — `objects[i]` is the 1-based id of the chained object that consumed `segments[i]` (0 means "didn't chain into any object", e.g. an isolated point). The frontend uses this to do object-level selection in 2D.
          items:
            format: uint32
            minimum: 0.0
            type: integer
          type: array
        segments:
          items:
            $ref: '#/components/schemas/Segment'
          type: array
        text_entities:
          description: DXF TEXT / MTEXT entities — emitted as editable metadata instead of being rendered to opaque polylines at import time. The frontend converts each into a `TextLayer` so the user can edit the content, font, size, rotation, etc., and the pipeline re-renders them at Generate. Empty for formats that don't carry text entities (SVG).
          items:
            $ref: '#/components/schemas/ImportedTextEntity'
          type: array
        unit_scale:
          format: double
          type: number
        warnings:
          items:
            type: string
          type: array
      required:
      - bbox
      - filename
      - format
      - layers
      - segments
      - unit_scale
      - warnings
      type: object
    ImportedObject:
      properties:
        bbox:
          $ref: '#/components/schemas/BBox'
        closed:
          type: boolean
        color:
          format: int32
          type: integer
        id:
          format: uint32
          minimum: 0.0
          type: integer
        layer:
          type: string
      required:
      - bbox
      - closed
      - color
      - id
      - layer
      type: object
    ImportedTextEntity:
      description: Metadata for a TEXT / MTEXT entity captured during import. Carries just enough for the frontend to construct a `TextLayer` — the editable inputs the pipeline pre-pass uses to render glyphs.
      properties:
        kind:
          $ref: '#/components/schemas/ImportedTextKind'
        origin:
          description: Anchor in stock XY (mm).
          items:
          - format: double
            type: number
          - format: double
            type: number
          maxItems: 2
          minItems: 2
          type: array
        rotation_deg:
          default: 0.0
          description: CCW degrees around `origin`.
          format: double
          type: number
        size_mm:
          format: double
          type: number
        source_layer:
          description: Original DXF layer name — preserved so the user knows where the text came from (the synthetic `__text_<id>` name takes over at pipeline time).
          type: string
        text:
          description: Full string. For Mtext, lines are `\n`-separated.
          type: string
      required:
      - kind
      - origin
      - size_mm
      - source_layer
      - text
      type: object
    ImportedTextKind:
      enum:
      - TEXT
      - MTEXT
      type: string
    Layer:
      properties:
        color:
          format: int32
          type: integer
        name:
          type: string
        segment_count:
          format: uint
          minimum: 0.0
          type: integer
      required:
      - color
      - name
      - segment_count
      type: object
...
    PostProcessorKind:
      enum:
      - linuxcnc
      - grbl
      - hpgl
      type: string
...
    RenderTextRequest:
      description: Request payload for the cross-transport `/text` endpoint. The frontend hands us TTF bytes (uploaded by the user or pulled from `frontend/public/fonts/`) plus a string + placement parameters; we return flattened [`Segment`]s and a single-line / outline classification the dialog uses to drive the engraving warning chip.
      properties:
        color:
          default: 7
          format: int32
          type: integer
        font_bytes:
          description: The font file as bytes (TTF / OTF). Encoded as a JSON `Vec<u8>` (i.e. an array of byte values) so the contract works equally well over HTTP/Tauri/WASM without picking a base64 layer.
          items:
            format: uint8
            minimum: 0.0
            type: integer
          type: array
        height_mm:
          format: double
          type: number
        layer:
          default: TEXT
          type: string
        origin:
          $ref: '#/components/schemas/Point2'
        text:
          type: string
      required:
      - font_bytes
      - height_mm
      - origin
      - text
      type: object
    RenderTextResponse:
      description: Response payload — the rendered geometry plus metadata the dialog uses to warn the user when an outline font is paired with the Engraving style.
      properties:
        family_name:
          description: Family / style names (best-effort). Useful for showing what the user actually loaded next to the dropdown.
          type:
          - string
          - 'null'
        segments:
          items:
            $ref: '#/components/schemas/Segment'
          type: array
        single_line:
          description: True if the font is a single-line / engraving / Hershey-port style font. Drives the dialog's "use a single-line font" chip.
          type: boolean
      required:
      - segments
      - single_line
      type: object
...
    Segment:
      description: |-
        A flat LINE/ARC primitive. ARC geometry is encoded as the bulge between `start` and `end` (bulge = `tan(included_angle / 4)`).

        `layer` is `Arc<str>` rather than `String` (jzpl Phase 2). A typical DXF has thousands of segments across a handful of layer names — the Arc lets every segment on the same layer share one allocation. Clone becomes a refcount bump; serde + `JsonSchema` treat it as a normal string on the wire.
      properties:
        bulge:
          default: 0.0
          format: double
          type: number
        center:
          anyOf:
          - $ref: '#/components/schemas/Point2'
          - type: 'null'
        color:
          default: 7
          format: int32
          type: integer
        end:
          $ref: '#/components/schemas/Point2'
        layer:
          default: '0'
          type: string
        start:
          $ref: '#/components/schemas/Point2'
        type:
          $ref: '#/components/schemas/SegmentKind'
      required:
      - end
      - start
      - type
      type: object
    SegmentKind:
      enum:
      - LINE
      - ARC
      - CIRCLE
      - POINT
      type: string
...
  responses:
    BadRequest:
      description: Malformed input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnprocessableEntity:
      description: Could not parse content
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'